update: reafctor lexer and make it simpler
previous version was ignoring the most of invalid tokens and i've realized it won't work if lexer ignore errors.
now everything proceed to be token but with "Invalid" type in error case

fix issue with numbers that start with zeros and.
fix issue with double numbers ending with dots.
fix issue with string literals that don't have closing quotes character